move gsl_ran_discrete_t_ptr into main namespace from internal namespace. Fix missing include of GSLrng_t.hpp in Makefile.am
