+++ fwdpp/Makefile.in
@@ -331,7 +331,9 @@ pkginclude_HEADERS = debug.hpp \
 	mlocuspop.hpp \
 	popgenmut.hpp \
 	simparams.hpp \
-	wrapped_range.hpp
+	wrapped_range.hpp \
+	GSLrng_t.hpp \
+	gsl_discrete.hpp
 
 all: all-recursive
