+++ fwdpp/gsl_discrete.hpp
@@ -6,11 +6,11 @@
 
 namespace fwdpp
 {
-    namespace fwdpp_internal
+    namespace detail
     {
         /*!
           Custom deleter for std::uniq_ptr
-        */
+		*/
         struct gsl_ran_discrete_t_deleter
         {
             void
@@ -19,14 +19,14 @@ namespace fwdpp
                 gsl_ran_discrete_free(l);
             }
         };
+    } // namespace detail
 
-        /*!
-          \warning Can only be put into a vector by push_back/emplace back
-          because of constraints on std::unique_ptr assignment
-         */
-        using gsl_ran_discrete_t_ptr
-            = std::unique_ptr<gsl_ran_discrete_t, gsl_ran_discrete_t_deleter>;
-    }
-}
+    /*!
+	  \warning Can only be put into a vector by push_back/emplace back
+	  because of constraints on std::unique_ptr assignment
+	 */
+    using gsl_ran_discrete_t_ptr
+        = std::unique_ptr<gsl_ran_discrete_t, detail::gsl_ran_discrete_t_deleter>;
+} // namespace fwdpp
 
 #endif
