+++ fwdpp/internal/Makefile.am
@@ -3,7 +3,6 @@ pkgincludedir=$(prefix)/include/fwdpp/internal
 pkginclude_HEADERS=mutation_internal.hpp \
 	rec_gamete_updater.hpp \
 	recombination_common.hpp \
-	gsl_discrete.hpp \
 	multilocus_rec.hpp \
 	ms_sampling.hpp \
 	gamete_cleaner.hpp \
