+++ fwdpp/internal/Makefile.in
@@ -268,7 +268,6 @@ top_srcdir = @top_srcdir@
 pkginclude_HEADERS = mutation_internal.hpp \
 	rec_gamete_updater.hpp \
 	recombination_common.hpp \
-	gsl_discrete.hpp \
 	multilocus_rec.hpp \
 	ms_sampling.hpp \
 	gamete_cleaner.hpp \
